+++ both/controllers/UserController.jsx
@@ -21,6 +21,12 @@ Users.actions({
     authentication:true,//AuthHelpers.userIsManagerofMembersTeam,
     method:RBAC.lib.destroy.bind(Users)
   },
+  getTeam:{
+    authentication:true,
+    helper:function(){
+      return Session.getSelectedTeam();
+    }
+  },
   getTeams:{
     authentication:true,
     helper:function(user){
@@ -28,6 +34,15 @@ Users.actions({
         {"members._id":user._id},
         {"owner._id":user._id}
       ]}).fetch()
+    }
+  },
+  getFacilities:{
+    authentication:true,
+    helper:function(user,q) {
+      var team = user.getTeam();
+      if(team) {
+        return team.getFacilities(q);
+      }
     }
   },
   getRole:{
@@ -51,6 +66,12 @@ Users.actions({
     authentication:true,
     helper:function(user) {
       return Session.getSelectedTeam()
+    }
+  },
+  getSelectedFacility:{
+    authentication:true,
+    helper:function(user) {
+      return Session.getSelectedFacility()
     }
   },
   getRequests:{
@@ -268,6 +289,9 @@ if(Meteor.isClient) {
       if(team) {
         Session.set('selectedTeam',{_id:team._id});
       }
+      else {
+        Session.set('selectedTeam',0);
+      }
       Session.set('selectedFacility',0);
     }
   Session.getSelectedFacility = function() {
@@ -279,6 +303,9 @@ if(Meteor.isClient) {
   Session.selectFacility = function(f) {
       if(f) {
         Session.set('selectedFacility',{_id:f._id});
+      }
+      else {
+        Session.set('selectedFacility',0);
       }
     }
   Session.getSelectedClient = function() {