+++ client/pages/ReportsPageIndex.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-import {GridView} from "react-gridview";
-
-ReportsIndexPage = React.createClass({
-	render() {
-		return <GridView/>
-	}
-})