+++ packages/fmc-autoform/inputs/MDSelect.jsx
@@ -11,6 +11,20 @@ PlainCard = React.createClass({
 })
 
 
+NameCard = React.createClass({
+	render() {
+		var name;
+		var item = this.props.item;
+		if(item) {
+			name = item.name;
+		}
+		return (
+			<span>{name}</span>
+		)
+	}
+})
+
+
 AutoInput.MDSelect = React.createClass({
 
 	getInitialState() {
@@ -83,7 +97,7 @@ AutoInput.MDSelect = React.createClass({
                 	>
 	                	{selectedItem?<Card item={selectedItem}/>:<span>&nbsp;</span>}
 	                </span>
-	                <div className="close-button" onClick={this.clearItem}>&times;</div>
+	                {selectedItem?<div className="close-button" onClick={this.clearItem}>&times;</div>:null}
 				    <span className="highlight"></span>
       				<span className="bar"></span>
                     <label>{this.props.placeholder}</label>