+++ packages/fmc-autoform/inputs/MDText.jsx
@@ -4,8 +4,16 @@ import {ReactMeteorData} from 'meteor/react-meteor-data';
 
 AutoInput.mdtext = React.createClass({
 
+	handleSelect(event) {
+		if(this.props.onSelect) {
+			this.props.onSelect(event.target.value);
+		}
+	},
+
 	handleChange(event) {
-		this.props.onChange(event.target.value);
+		if(this.props.onChange) {
+			this.props.onChange(event.target.value);
+		}
 	},
 
 	componentDidMount() {
@@ -28,6 +36,13 @@ AutoInput.mdtext = React.createClass({
 		}
 	},
 
+	handleClearItem() {
+		if(this.props.onClear) {
+			this.props.onClear()
+		}
+		this.handleChange({target:{value:""}});
+	},
+
 	render() {
 		var value, used;
 		value = this.props.value||"";
@@ -43,8 +58,10 @@ AutoInput.mdtext = React.createClass({
       			pattern=".{1,80}" 
       			className={"input "+(used?'used':'')} 
       			value={value}
+      			onSelect={this.handleSelect}
       			onChange={this.handleChange}
       		/>
+	        {used&&value.length?<div className="close-button" onClick={this.handleClearItem}>&times;</div>:null}
       		<span className="highlight"></span>
       		<span className="bar"></span>
       		<label>{this.props.placeholder}</label>