+++ CHANGELOG.md
@@ -1,6 +1,49 @@
 CHANGELOG.md
 ============
 
+**TODO**
+
+* Fixed header scrolling table
+
+* Export flag needs to be set when an item is exported (...or if not then when)
+* Right align numbers
+
+* Link to report from widgets on dashboard
+* Abstract toolbar from dash and report (and perhaps fab can inherit same mechanism - might be flux pattern)
+
+* Need a way to stack actions into toolbar for reports (fab will follow this same mechanism)
+* Need "card" navigation view for reports (filterbox 2 to be generalised?)
+
+* Problem - there is no way to set the role of users who are not "owned" by you or your team
+* Problem - role confusion (manager perhaps should be facility manager)
+* Problem - setting rbac is confusing
+* Problem - Managers can add personnel but do not have permission to edit those owned by team
+
+* Need to be able to clear text (particularly for new date picker)
+
+**v1.4.3b**
+
+* Fixed a problem that was preventing reports from exporting correctly
+
+
+**v1.4.2b**
+
+* Links to reports go on dashboard
+* Only show updates on dashboard if allowed to see
+* Removed export button from fab
+* Filter unit that goes at top of reports
+* New datetime picker
+* Formatting for int and float values in datagrid
+ 
+**v1.4.1b**
+
+* Added requests status report
+* Created new flexible md compliant data grid view
+* CSV export for reports
+* Fixed a problem that was causing users to not be saved by "managers"
+* Added report module
+* Dashboard elements can now be viewed as reports and vice versa
+
 **v1.4.0b**
 
 * Restricted manager role so that manager only see facilities they are a member of